Extract utility functions
